+++ nf_core/schema.py
@@ -5,19 +5,15 @@ from __future__ import print_function
 from rich.prompt import Confirm
 
 import copy
+import copy
 import jinja2
 import json
 import jsonschema
 import logging
+import markdown
 import os
-import requests
-import requests_cache
-import sys
-import time
 import webbrowser
 import yaml
-import copy
-import re
 
 import nf_core.list, nf_core.utils
 
@@ -415,6 +411,91 @@ class PipelineSchema(object):
             assert self.schema["description"] == desc_attr, "Schema 'description' should be '{}'\n Found: '{}'".format(
                 desc_attr, self.schema["description"]
             )
+
+    def print_documentation(
+        self,
+        output_fn=None,
+        format="markdown",
+        force=False,
+        columns=["parameter", "description", "type,", "default", "required", "hidden"],
+    ):
+        """
+        Prints documentation for the schema.
+        """
+        output = self.schema_to_markdown(columns)
+        if format == "html":
+            output = self.markdown_to_html(output)
+
+        # Print to file
+        if output_fn:
+            if os.path.exists(output_fn) and not force:
+                log.error(f"File '{output_fn}' exists! Please delete first, or use '--force'")
+                return
+            with open(output_fn, "w") as file:
+                file.write(output)
+                log.info(f"Documentation written to '{output_fn}'")
+
+        # Print to stdout
+        else:
+            print(output)
+
+    def schema_to_markdown(self, columns):
+        """
+        Creates documentation for the schema in Markdown format.
+        """
+        out = f"# {self.schema['title']}\n\n"
+        out += f"{self.schema['description']}\n"
+        # Grouped parameters
+        for definition in self.schema.get("definitions", {}).values():
+            out += f"\n## {definition.get('title', {})}\n\n"
+            out += f"{definition.get('description', '')}\n\n"
+            out += "".join([f"| {column.title()} " for column in columns])
+            out += "|\n"
+            out += "".join([f"|-----------" for columns in columns])
+            out += "|\n"
+            for p_key, param in definition.get("properties", {}).items():
+                for column in columns:
+                    if column == "parameter":
+                        out += f"| `{p_key}` "
+                    elif column == "description":
+                        out += f"| {param.get('description', '')} "
+                        if param.get("help_text", "") != "":
+                            out += f"<details><summary>Help</summary><small>{param['help_text']}</small></details>"
+                    elif column == "type":
+                        out += f"| `{param.get('type', '')}` "
+                    else:
+                        out += f"| {param.get(column, '')} "
+                out += "|\n"
+
+        # Top-level ungrouped parameters
+        if len(self.schema.get("properties", {})) > 0:
+            out += f"\n## Other parameters\n\n"
+            out += "".join([f"| {column.title()} " for column in columns])
+            out += "|\n"
+            out += "".join([f"|-----------" for columns in columns])
+            out += "|\n"
+
+            for p_key, param in self.schema.get("properties", {}).items():
+                for column in columns:
+                    if column == "parameter":
+                        out += f"| `{p_key}` "
+                    elif column == "description":
+                        out += f"| {param.get('description', '')} "
+                        if param.get("help_text", "") != "":
+                            out += f"<details><summary>Help</summary><small>{param['help_text']}</small></details>"
+                    elif column == "type":
+                        out += f"| `{param.get('type', '')}` "
+                    else:
+                        out += f"| {param.get(column, '')} "
+                out += "|\n"
+
+            return out
+
+    def markdown_to_html(self, markdown_str):
+        """
+        Convert markdown to html
+        """
+        return markdown.markdown(markdown_str, extensions=["tables"])
 
     def make_skeleton_schema(self):
         """Make a new pipeline schema from the template"""